+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 1. [](#new)
     * Set minimum requirements to [PHP 5.6.4](https://getgrav.org/blog/raising-php-requirements-2018)
     * Updated Doctrine Collections to 1.4
-    * Updated Symfony Components to 3.4, causing some compatibility issues with YAML files
+    * Updated Symfony Components to 3.4 (with compatibility mode to fall back to Symfony YAML 2.8)
 
 # v1.4.4
 ## 04/12/2018